# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: FIS Process API
  version: v1
  description: Complete REST API contract for the FIS ledger service.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /reports/trial-balance:
    get:
      summary: Generate trial balance report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: asOfDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Trial balance report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrialBalanceReport'
  /reports/balance-sheet:
    get:
      summary: Generate balance sheet report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: asOfDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Balance sheet report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceSheetReport'
  /reports/income-statement:
    get:
      summary: Generate income statement report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Income statement report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomeStatementReport'
  /reports/general-ledger/{accountCode}:
    get:
      summary: Generate general ledger detail report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: accountCode
          in: path
          required: true
          schema:
            type: string
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: General ledger report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralLedgerReport'
  /reports/cash-flow:
    get:
      summary: Generate cash flow report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Cash flow report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashFlowReport'
  /reports/account-activity/{accountCode}:
    get:
      summary: Generate account activity report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: accountCode
          in: path
          required: true
          schema:
            type: string
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Account activity report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountActivityReport'
  /reports/journal-register:
    get:
      summary: Generate journal register report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Journal register report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalRegisterReport'
  /reports/dimension-summary:
    get:
      summary: Generate dimension summary report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: dimensionKey
          in: query
          required: true
          schema:
            type: string
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Dimension summary report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DimensionSummaryReport'
  /reports/fx-exposure:
    get:
      summary: Generate FX exposure report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: asOfDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: FX exposure report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxExposureReport'
  /reports/aging:
    get:
      summary: Generate aging analysis report
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: accountType
          in: query
          required: true
          schema:
            type: string
            enum: [ASSET, LIABILITY]
        - name: asOfDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Aging report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgingReport'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ReportMetadata:
      type: object
      properties:
        reportType:
          type: string
        tenantId:
          type: string
          format: uuid
        asOfDate:
          type: string
          format: date
          nullable: true
        fromDate:
          type: string
          format: date
          nullable: true
        toDate:
          type: string
          format: date
          nullable: true
        baseCurrency:
          type: string
        generatedAt:
          type: string
          format: date-time
    TrialBalanceLine:
      type: object
      properties:
        accountCode:
          type: string
        accountName:
          type: string
        accountType:
          type: string
        debitBalance:
          type: integer
          format: int64
        creditBalance:
          type: integer
          format: int64
    TrialBalanceReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        lines:
          type: array
          items:
            $ref: '#/components/schemas/TrialBalanceLine'
        totalDebits:
          type: integer
          format: int64
        totalCredits:
          type: integer
          format: int64
        balanced:
          type: boolean
    BalanceSheetLine:
      type: object
      properties:
        accountCode:
          type: string
        accountName:
          type: string
        balanceCents:
          type: integer
          format: int64
        formattedBalance:
          type: string
    BalanceSheetSection:
      type: object
      properties:
        sectionName:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/BalanceSheetLine'
        sectionTotal:
          type: integer
          format: int64
    BalanceSheetReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        assets:
          $ref: '#/components/schemas/BalanceSheetSection'
        liabilities:
          $ref: '#/components/schemas/BalanceSheetSection'
        equity:
          $ref: '#/components/schemas/BalanceSheetSection'
        totalAssets:
          type: integer
          format: int64
        totalLiabilitiesAndEquity:
          type: integer
          format: int64
        balanced:
          type: boolean
    IncomeStatementLine:
      type: object
      properties:
        accountCode:
          type: string
        accountName:
          type: string
        amountCents:
          type: integer
          format: int64
        formattedAmount:
          type: string
    IncomeStatementSection:
      type: object
      properties:
        sectionName:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/IncomeStatementLine'
        sectionTotal:
          type: integer
          format: int64
    IncomeStatementReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        revenue:
          $ref: '#/components/schemas/IncomeStatementSection'
        expenses:
          $ref: '#/components/schemas/IncomeStatementSection'
        totalRevenue:
          type: integer
          format: int64
        totalExpenses:
          type: integer
          format: int64
        netIncome:
          type: integer
          format: int64
    GeneralLedgerEntry:
      type: object
      properties:
        journalEntryId:
          type: string
          format: uuid
        sequenceNumber:
          type: integer
          format: int64
          nullable: true
        postedDate:
          type: string
          format: date
        description:
          type: string
        debitAmount:
          type: integer
          format: int64
        creditAmount:
          type: integer
          format: int64
        runningBalance:
          type: integer
          format: int64
    GeneralLedgerReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        accountCode:
          type: string
        accountName:
          type: string
        openingBalance:
          type: integer
          format: int64
        entries:
          type: array
          items:
            $ref: '#/components/schemas/GeneralLedgerEntry'
        closingBalance:
          type: integer
          format: int64
    CashFlowLine:
      type: object
      properties:
        label:
          type: string
        amountCents:
          type: integer
          format: int64
    CashFlowSection:
      type: object
      properties:
        sectionName:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/CashFlowLine'
        sectionTotal:
          type: integer
          format: int64
    CashFlowReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        operatingActivities:
          $ref: '#/components/schemas/CashFlowSection'
        investingActivities:
          $ref: '#/components/schemas/CashFlowSection'
        financingActivities:
          $ref: '#/components/schemas/CashFlowSection'
        netCashChange:
          type: integer
          format: int64
        openingCash:
          type: integer
          format: int64
        closingCash:
          type: integer
          format: int64
    AccountActivityReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        accountCode:
          type: string
        accountName:
          type: string
        openingBalance:
          type: integer
          format: int64
        totalDebits:
          type: integer
          format: int64
        totalCredits:
          type: integer
          format: int64
        closingBalance:
          type: integer
          format: int64
        transactionCount:
          type: integer
          format: int64
    JournalRegisterEntry:
      type: object
      properties:
        journalEntryId:
          type: string
          format: uuid
        sequenceNumber:
          type: integer
          format: int64
          nullable: true
        postedDate:
          type: string
          format: date
        description:
          type: string
        status:
          type: string
        totalDebits:
          type: integer
          format: int64
        totalCredits:
          type: integer
          format: int64
        createdBy:
          type: string
    JournalRegisterReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/JournalRegisterEntry'
        totalEntries:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
          format: int64
    DimensionSummaryLine:
      type: object
      properties:
        dimensionValue:
          type: string
        debitTotal:
          type: integer
          format: int64
        creditTotal:
          type: integer
          format: int64
        netAmount:
          type: integer
          format: int64
    DimensionSummaryReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        dimensionKey:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/DimensionSummaryLine'
        grandTotal:
          type: integer
          format: int64
    FxExposureLine:
      type: object
      properties:
        currency:
          type: string
        assetExposure:
          type: integer
          format: int64
        liabilityExposure:
          type: integer
          format: int64
        netExposure:
          type: integer
          format: int64
        currentRate:
          type: number
        unrealizedGainLoss:
          type: integer
          format: int64
    FxExposureReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        baseCurrency:
          type: string
        exposures:
          type: array
          items:
            $ref: '#/components/schemas/FxExposureLine'
        totalUnrealizedGainLoss:
          type: integer
          format: int64
    AgingBucket:
      type: object
      properties:
        bucketLabel:
          type: string
        amountCents:
          type: integer
          format: int64
        entryCount:
          type: integer
          format: int64
    AgingReport:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/ReportMetadata'
        accountType:
          type: string
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/AgingBucket'
        grandTotal:
          type: integer
          format: int64